+++ orderTest/addons/IsFunctions.cs
@@ -11,67 +11,42 @@ namespace orderTest
     public partial class Form1 : Form
     {
         //all
-        private bool isLast(int count) { if (count > 0) { return true; } return false; }
+        //private bool isLast(int count) { if (count > 0) { return true; } return false; }
 
-        private void isNullPosition(ComboBox a, Control b) { if (a.SelectedIndex != 0) { b.Enabled = true; } else { b.Enabled = false; } }
+        //private void isNullPosition(ComboBox a, Control b) { if (a.SelectedIndex != 0) { b.Enabled = true; } else { b.Enabled = false; } }
 
-        private void selectDefaultItem(ComboBox[] cb) { foreach (ComboBox cbItem in cb) { cbItem.SelectedIndex = 0; } }
+        //private void selectDefaultItem(ComboBox[] cb) { foreach (ComboBox cbItem in cb) { cbItem.SelectedIndex = 0; } }
 
-        private void fillEnable(Control[] ctrs, bool state, string type = "t") { foreach (Control c in ctrs) { c.Enabled = state; } }
-
-        private void fillEnable(List<Control> ctrs, bool state, string type = "t") { foreach (Control c in ctrs) { c.Enabled = state; } }
+        //private void fillEnable(Control[] ctrs, bool state, string type = "t") { foreach (Control c in ctrs) { c.Enabled = state; } }
 
         //реквізити
-        private void txt(List<TextBox> list, string[] ph) { foreach (TextBox t in list) { t.Text = "---" + ph[list.IndexOf(t)] + "---"; } }
+        //private void txt(List<Control> list, string[] ph) { foreach (TextBox t in list) { t.Text = "---" + ph[list.IndexOf(t)] + "---"; } }
 
-        private bool isPH(string ph) { if (ph.Contains("---")) { return true; } return false; }
+        //private bool isPH(string ph) { if (ph.Contains("---")) { return true; } return false; }
 
-        private void clearText(TextBox tb) { if(isPH(tb.Text)) { tb.Clear(); } }
+        //private void clearText(TextBox tb) { if(isPH(tb.Text)) { tb.Clear(); } }
 
-        private void isFill(TextBox t, Control[] ctrs, bool state, string ph) { if (!t.Text.Any()) { txt(new List<TextBox> { t }, new[] { ph }); fillEnable(ctrs, state); } }
+        //private void isFill(TextBox t, Control[] ctrs, bool state, string ph) { if (!t.Text.Any()) { txt(new List<Control> { t }, new[] { ph }); fillEnable(ctrs, state); } }
 
-        private void headClear(Control[] ca)
-        {
-            int count = 0;
-            MessageBox.Show(count.ToString());
-            List<Control> controls = ca.ToList();
-            
-            count = controls.Count();
-            MessageBox.Show(count.ToString());
-            _ = controls.Append(numberHead);//.Append<Control>(clearHead);
-            count = controls.Count();
-            MessageBox.Show(count.ToString());
-            //Control[] ctrs = (Control[])tbArray.Concat<Control[]>(new Control[] { numberHead });
-            //tbArray = new Control[] { markHead, vehicleHead, addressHead, driverHead };
-            //Control[] ctrs2 = (Control[])ctrs1.Concat(tbArray);
-            dateHead.ResetText();
-            txt(new List<TextBox> { numberHead, clientHead, markHead, vehicleHead, addressHead, driverHead }, new[] { "#", "замовник", "марка", "машина", "адреса", "водій" });
-            fillEnable(controls, false);
-            count = controls.Count();
-            MessageBox.Show(count.ToString());
-            //fillEnable(new[] { clearHead, addHeadData }, false, "b");
-        }
+        //private void headClear(List<Control> cl)
+        //{
+        //    List<Control> tmpCtrTxt = new List<Control> { numberHead, clientHead }.Concat(cl).ToList();
+        //    dateHead.ResetText();
+        //    txt(tmpCtrTxt, new[] { "#", "замовник", "марка", "машина", "адреса", "водій" }); fillEnable(tmpCtrTxt.Concat(new List<Control> { clearHead, addHeadData }).ToArray(), false);
+        //}
 
         //пінопласт
 
-        private void isMark(string[] row) { if (EpsList.Last().Mark.Equals(markEPS.Text)) { row[0] = ""; } }
+        //private void isMark(string[] row) { if (EpsList.Last().Mark.Equals(markEPS.Text)) { row[0] = ""; } }
 
-        private void epsClear()
-        {
-            markEPS.SelectedIndex = 0;
-            thikEPS.SelectedIndex = 0; thikEPS.Enabled = false;
-            amountEPS.Clear(); amountEPS.Enabled = false;
-            packEPS.Clear(); packEPS.Enabled = false;
+        //private void epsClear()
+        //{
+        //    selectDefaultItem(new[] { markEPS, thikEPS }); amountEPS.Clear(); packEPS.Clear(); fillEnable(new Control[] { thikEPS, amountEPS, packEPS, addEpsToOrderButton }, false);
+        //}
 
-            addEpsToOrderButton.Enabled = false;
-        }
-
-        private double div(string s) { if (s.Equals("8")) { return 0.32; }; return 0.3; }
+        //private double div(string s) { if (s.Equals("8")) { return 0.32; }; return 0.3; }
 
         //додаткові
-        private void addClear()
-        {
-            nameADD.SelectedIndex = 0; meterADD.SelectedIndex = 0; meterADD.Enabled = false; amountADD.Clear(); amountADD.Enabled = false; addAddToOrderButton.Enabled = false;
-        }
+        //private void addClear() { selectDefaultItem(new[] { nameADD, meterADD }); amountADD.Clear(); fillEnable(new Control[] { meterADD, amountADD, addAddToOrderButton }, false); }
     }
 }
