+++ orderTest/orderTest.csproj
@@ -48,6 +48,18 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="addons\AddFunction.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="addons\EpsFunction.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="addons\GenFunction.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="addons\HeadFunction.cs">
+      <SubType>Form</SubType>
+    </Compile>
     <Compile Include="addons\IsFunctions.cs">
       <SubType>Form</SubType>
     </Compile>
